fix: update unit tests for data gen changes
