+++ tests/unit/internal/models/execution/test_process_executor.py
@@ -33,12 +33,12 @@ class TestProcessExecutor:
         self.code = "print('Hello, World!')"
         self.working_dir = Path(os.getcwd()) / self.execution_id
         self.timeout = 5
-        self.dataset = pd.DataFrame({"col1": [1, 2], "col2": [3, 4]})
+        self.datasets = {"training_data": pd.DataFrame({"col1": [1, 2], "col2": [3, 4]})}
         self.process_executor = ProcessExecutor(
             execution_id=self.execution_id,
             code=self.code,
             working_dir=Path(os.getcwd()),
-            dataset=self.dataset,
+            datasets=self.datasets,
             timeout=self.timeout,
             code_execution_file_name="run.py",
         )
@@ -60,7 +60,7 @@ class TestProcessExecutor:
             result = self.process_executor.run()
 
         dataset_file = self.working_dir / "training_data.parquet"
-        mock_write_table.assert_called_once_with(pyarrow.Table.from_pandas(self.dataset), dataset_file)
+        mock_write_table.assert_called_once_with(pyarrow.Table.from_pandas(self.datasets["training_data"]), dataset_file)
         mock_popen.assert_called_once_with(
             [sys.executable, str(self.working_dir / "run.py")],
             stdout=subprocess.PIPE,
@@ -101,7 +101,7 @@ class TestProcessExecutor:
     def test_dataset_written_to_file(self, mock_write_table):
         self.process_executor.run()
         dataset_file = self.working_dir / "training_data.parquet"
-        mock_write_table.assert_called_once_with(pyarrow.Table.from_pandas(self.dataset), dataset_file)
+        mock_write_table.assert_called_once_with(pyarrow.Table.from_pandas(self.datasets["training_data"]), dataset_file)
 
 
 if __name__ == "__main__":
